pep8 reformat code

+++ python/federatedml/nn/dataset/image.py
@@ -42,7 +42,8 @@ class ImageDataset(Dataset):
         self.float64 = float64
         self.dtype = torch.float32 if not self.float64 else torch.float64
         avail_label_type = ['float', 'long', 'double']
-        assert label_dtype in avail_label_type, 'available label dtype : {}'.format(avail_label_type)
+        assert label_dtype in avail_label_type, 'available label dtype : {}'.format(
+            avail_label_type)
         if label_dtype == 'double':
             self.label_dtype = torch.float64
         elif label_dtype == 'long':
@@ -54,7 +55,8 @@ class ImageDataset(Dataset):
 
         # read image from folders
         if self.center_crop:
-            transformer = transforms.Compose([transforms.CenterCrop(size=self.size), transforms.ToTensor()])
+            transformer = transforms.Compose(
+                [transforms.CenterCrop(size=self.size), transforms.ToTensor()])
         else:
             transformer = transforms.Compose([transforms.ToTensor()])
 
@@ -69,14 +71,18 @@ class ImageDataset(Dataset):
             file_name = self.image_folder.imgs
             ids = []
             for name in file_name:
-                sample_id = name[0].split('/')[-1].replace(self.file_suffix, '')
+                sample_id = name[0].split(
+                    '/')[-1].replace(self.file_suffix, '')
                 ids.append(sample_id)
             self.set_sample_ids(ids)
 
     def __getitem__(self, item):
         if self.return_label:
             item = self.image_folder[item]
-            return item[0].type(self.dtype), torch.tensor(item[1]).type(self.label_dtype)
+            return item[0].type(
+                self.dtype), torch.tensor(
+                item[1]).type(
+                self.label_dtype)
         else:
             return self.image_folder[item][0].type(self.dtype)
